Refactor code following new naming convention
I have decided to unify naming convention for all functions in the code,
so I will not distinguish between member and non-member functions.
Although I am not fan of prefixes, in C++ it makes sense to set up a
convention for member variable names.  I find 'mCamelCase' as the best
convention in respect to code readability.

The comments underwent minor improvements and I have decided to add
concise explanation for why it is necessary to include empty
argumentless constructor in the case when parametrized constructor is
defined.

Some local variables in the 'InventoryQueue' member functions also
underwent slight refactoring to make code more understandable.