Explicitly initialize transaction structure
It is good idea to have transaction initialized with zero values when
parsing the file.  In case if the transaction record was not parsed
properly, null values will guarantee that the invalid transaction will
not be applied to the queue (since zero values are considered as
invalid).  Without explicit initialization there is a chance that those
variables will contain random values and thus produce unpredictable
results.